rearranged app to separate receiver / sender for ease of communicaton across function apps
