+++ 01_serverless_architecture/serverless_app/azure_host_receiver/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "azure_host_receiver": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7151",
+      "launchBrowser": false
+    }
+  }
+}
