+++ 01_serverless_architecture/serverless_app/azure_host_receiver/host.json
@@ -1,0 +1,5 @@
+{
+    "version": "2.0",
+    "logging": {
+    }
+}
