+++ 01_serverless_architecture/serverless_app/azure_host_sender/Properties/launchSettings.json
@@ -1,0 +1,9 @@
+{
+  "profiles": {
+    "azure_host_sender": {
+      "commandName": "Project",
+      "commandLineArgs": "--port 7284",
+      "launchBrowser": false
+    }
+  }
+}
