+++ 01_serverless_architecture/serverless_app/gcp_host_receiver/gcp_host_receiver.csproj
@@ -5,12 +5,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Google.Cloud.Functions.Hosting" Version="1.1.0" />
+    <PackageReference Include="Google.Cloud.Functions.Hosting" Version="2.0.0" />
     <None Include="appsettings*.json" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\sender\sender.csproj" />
     <ProjectReference Include="..\receiver\receiver.csproj" />
     <ProjectReference Include="..\data\data.csproj" />
   </ItemGroup>
