+++ 01_serverless_architecture/serverless_app/gcp_host_sender/appsettings.json
@@ -1,0 +1,3 @@
+{
+    "RECEIVERADDR": "http://localhost:7071/api/sender"
+}
