+++ 01_serverless_architecture/terraform/providers.tf
@@ -11,12 +11,15 @@ terraform {
   }
 
   # location of remote state (this is independent of accompanying resources. e.g. backend could be AWS or local)
-  backend "azurerm" {
-    # resource_group_name  = "state_resource_group"
-    # storage_account_name = "state_storage_account"
-    # container_name       = "state_container"
-    # key                  = "state_filename"
+  backend "local" {
+    path = "serverless_app.tfstate"
   }
+  # backend "azurerm" {
+  #   # resource_group_name  = "state_resource_group"
+  #   # storage_account_name = "state_storage_account"
+  #   # container_name       = "state_container"
+  #   # key                  = "state_filename"
+  # }
   # initialising secretly, without configuring directly within providers.tf:
   # terraform init -backend-config="resource_group_name=<state_resource_group>" \
   #   -backend-config="storage_account_name=<state_storage_account>" \
@@ -31,5 +34,5 @@ provider "azurerm" {
 provider "google" {
   project = var.google_provider_project
   region  = var.google_provider_region
-  zone    = var.google_provider_zone
+#   zone    = var.google_provider_zone
 }
